feat(admin): simplify dashboard and add feature links
