+++ app/admin/ai-tools/page.tsx
@@ -1,0 +1,8 @@
+export default function AiToolsPage() {
+  return (
+    <div className="p-4 space-y-2">
+      <h1 className="text-2xl font-bold">AI Tools</h1>
+      <p>หน้านี้ยังเป็น mock</p>
+    </div>
+  )
+}
