+++ app/admin/feature-map/page.tsx
@@ -1,0 +1,8 @@
+export default function FeatureMapPage() {
+  return (
+    <div className="p-4 space-y-2">
+      <h1 className="text-2xl font-bold">Feature Map</h1>
+      <p>หน้านี้เป็นตัวอย่างแผนที่ฟีเจอร์</p>
+    </div>
+  )
+}
