+++ public/mock/chat.json
@@ -1,0 +1,7 @@
+[
+  {"id": "C1", "createdAt": "2024-06-08T09:00:00Z", "text": "สวัสดี"},
+  {"id": "C2", "createdAt": "2024-06-08T12:00:00Z", "text": "สอบถามราคา"},
+  {"id": "C3", "createdAt": "2024-06-07T14:30:00Z", "text": "มีสีอื่นไหม"},
+  {"id": "C4", "createdAt": "2024-06-06T10:15:00Z", "text": "ขอดูรูปเพิ่มเติม"},
+  {"id": "C5", "createdAt": "2024-06-05T16:45:00Z", "text": "ขอบคุณค่ะ"}
+]
